% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcipSupplementalLab-package.R
\docType{package}
\name{fcipSupplementalLab-package}
\alias{fcipSupplementalLab}
\alias{fcipSupplementalLab-package}
\title{fcipSupplementalLab: Research Framework for Supplemental Crop Insurance in the FCIP}
\description{
Provides a research framework for analyzing supplemental crop insurance products in the United States Federal Crop Insurance Program (FCIP). The package supports reproducible workflows to evaluate adoption and demand, actuarial performance and program soundness, fiscal exposure, risk reduction and income transfer, and basis risk and coverage quality. Functions emphasize transparent data preparation, diagnostic summaries, and modular analysis components suitable for reports and policy briefs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ftsiboe/fcipSupplementalLab}
  \item Report bugs at \url{https://github.com/ftsiboe/fcipSupplementalLab/issues}
}

}
\keyword{internal}
