% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcipSupplementalLab_controls.R
\name{fcipSupplementalLab_controls}
\alias{fcipSupplementalLab_controls}
\title{Create a control list of adjustment factors for FCIP-related packages}
\usage{
fcipSupplementalLab_controls()
}
\value{
A named list of control parameters, ready to be passed to other simulation functions.
}
\description{
Create a control list of adjustment factors for FCIP-related packages
}
\concept{helpers}
